refactor backend code to MVC model
